Add documentation to Trie and TrieNode (Separated) Hijacked "tab" key in command box for auto completion
