+++ src/main/java/seedu/address/model/datastructure/Trie.java
@@ -1,30 +1,41 @@
-package seedu.address.model;
+package seedu.address.model.datastructure;
 
 import java.util.Collection;
-import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Set;
 
 import org.apache.commons.lang.StringUtils;
 
+//@@author A0140042A
 /**
  * Data structure used to store auto complete data
  */
 public class Trie {
     private TrieNode rootNode;
 
+    /**
+     * Initializes a new Trie datastructure with an empty root node
+     */
     public Trie() {
         super();
         rootNode = new TrieNode(' ');
     }
 
+    /**
+     * Loads phrases used for auto completion into the Trie
+     * @param phrases
+     */
     public void load(String... phrases) {
         for (String phrase : phrases) {
             loadRecursive(rootNode, phrase + "$");
         }
     }
 
+    /**
+     * Loads phrases character by character into the Trie
+     * @param node - root to add the phrase from
+     * @param phrase - phrase to add
+     */
     private void loadRecursive(TrieNode node, String phrase) {
         if (StringUtils.isBlank(phrase)) {
             return;
@@ -37,11 +48,22 @@ public class Trie {
         }
     }
 
+    /**
+     * Checks if the Trie contains a specific prefix
+     * @param prefix - the string to match
+     * @return true if there is a match to prefix
+     */
     public boolean matchPrefix(String prefix) {
         TrieNode matchedNode = matchPrefixRecursive(rootNode, prefix);
         return (matchedNode != null);
     }
 
+    /**
+     * Checks recursively if each node contains a specific prefix
+     * @param node - The node to check
+     * @param prefix - the string to match
+     * @return true if the node or any of its children contains the prefix
+     */
     private TrieNode matchPrefixRecursive(TrieNode node, String prefix) {
         if (StringUtils.isBlank(prefix)) {
             return node;
@@ -57,6 +79,11 @@ public class Trie {
         }
     }
 
+    /**
+     * Finds a list of completions that can be made using the current prefix
+     * @param prefix - to match against
+     * @return a list of strings with auto completed suggestions
+     */
     public List<String> findCompletions(String prefix) {
         TrieNode matchedNode = matchPrefixRecursive(rootNode, prefix);
         List<String> completions = new LinkedList<String>();
@@ -64,6 +91,12 @@ public class Trie {
         return completions;
     }
 
+    /**
+     * Finds a list of compeletions that can be made using the current prefix recursively
+     * @param node - current node to check
+     * @param prefix - to match against
+     * @param completions - List of completions that are currently matched
+     */
     private void findCompletionsRecursive(TrieNode node, String prefix, List<String> completions) {
         if (node == null) {
             // our prefix did not match anything, just return
@@ -81,57 +114,5 @@ public class Trie {
             char childChar = childNode.getNodeValue();
             findCompletionsRecursive(childNode, prefix + childChar, completions);
         }
-    }
-}
-
-// TrieNode.java
-class TrieNode {
-
-    private Character character;
-    private HashMap<Character, TrieNode> children;
-
-    public TrieNode(char c) {
-        super();
-        this.character = new Character(c);
-        children = new HashMap<Character, TrieNode>();
-    }
-
-    public char getNodeValue() {
-        return character.charValue();
-    }
-
-    public Collection<TrieNode> getChildren() {
-        return children.values();
-    }
-
-    public Set<Character> getChildrenNodeValues() {
-        return children.keySet();
-    }
-
-    public void add(char c) {
-        if (children.get(new Character(c)) == null) {
-            // children does not contain c, add a TrieNode
-            children.put(new Character(c), new TrieNode(c));
-        }
-    }
-
-    public TrieNode getChildNode(char c) {
-        return children.get(new Character(c));
-    }
-
-    public boolean contains(char c) {
-        return (children.get(new Character(c)) != null);
-    }
-
-    public int hashCode() {
-        return character.hashCode();
-    }
-
-    public boolean equals(Object obj) {
-        if (!(obj instanceof TrieNode)) {
-            return false;
-        }
-        TrieNode that = (TrieNode) obj;
-        return (this.getNodeValue() == that.getNodeValue());
     }
 }
