+++ src/main/java/seedu/address/ui/CommandBox.java
@@ -5,6 +5,8 @@ import java.util.logging.Logger;
 import javafx.fxml.FXML;
 import javafx.scene.control.SplitPane;
 import javafx.scene.control.TextField;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.Region;
 import seedu.address.commons.core.LogsCenter;
@@ -56,6 +58,14 @@ public class CommandBox extends UiPart<Region> {
         }
     }
 
+    @FXML
+    private void handleOnKeyPressed(KeyEvent ke) {
+        if (ke.getCode() == KeyCode.TAB) {
+            String typed = commandTextField.getText();
+            System.out.println(typed);
+            ke.consume();
+        }
+    }
 
     /**
      * Sets the command box style to indicate a successful command.
